Use Solaris 10 compatible compiler triplet
This changes the solaris2.11 part of the gcc triplet
to solaris2.10, because there is a compatibility issue:
If a rust program panics on Solaris 10 the backtrace
cannot be generated because of a runtime error.

Changing the triplet works for both Solaris 10 and 11.
It is also the same triplet that is used by Rust's CI.