Use ssh to execute binaries in qemu-system
This avoids hacks to get the exit code of the executed programs and also
allows the correct color for the host terminal. It increases the time to
start the qemu-system emulator and to start the execution of a program
on it. But the overall time to execute ripgrep tests does not changes.

We use dropbear because the openssh client refuses to execute in docker
with an unknown user id.