Check for updates on base image during build
It is important to keep Docker images up-to date, as upstream images
could have important security fixes, even using the same tag and
version.

When building images, it is possible to pass in a `--pull` flag to check
if the base image has updates, and rebuild as necessary. This feature is
available since 2014, and will still retain the image cache if there are
no updates on the base image.

[since 2014]: https://github.com/moby/moby/pull/9281