Configure ports.ubuntu.com in APT sources
The recommendation given in the README to use `dpkg --add-architecture` to install additional libraries doesn't actually work because the packages for most architectures are on a different server.

This change configures the APT sources so it knows where every architecture is located, and the recommendation given in the README works.

The package locations should be accurate for all current Ubuntu releases, although I have only tested it with 16.04 and just looked at the repository content for later versions.

I have tested these two shell commands in a docker image based on the official cross docker images. I think this would be the best place to put them so that all cross users can benefit from them, and they *should* work here but I've not tried to build an image.