docker: linux-image: add gnupg to local deps
It's needed when generating an image based on Ubuntu 18.04 or Debian
Buster. Adding it makes it easier for users to create such image as they
can just re-use the existing script without having to modify it.
It doesn't matter when building with Ubuntu 16.04 as the package is
already present in the base image so it's just skipped.