Try to install only available toolchains via rustup
This prevents cross from erroring out and exiting when we have a custom
target which we know will be available in the docker image (but is not
one of the standard rustup targets). An example of such a target is
"x86_64-alpine-linux-musl" which is a custom target that the alpine
distro uses and is available when using rustc from its repositories.