Specify `edge` for the example of Dockerfile
`README.md` recommends writing `Dockerfile` based on the default
Docker image that cross uses. The example indicates the base
image's tag as `latest`, but `latest` does not seem to exist on container
packages.

https://github.com/orgs/cross-rs/packages?repo_name=cross

So, I updated `README.md` to specify `edge` instead of `latest`
in order for developers to use the snippet just by copying it.