Fix ARMv6 HF image to correctly build for ARMv6 architecture.
Change the `arm-unknown-linux-gnueabihf` to build a toolchain using crosstool-ng, so that we build a toolchain using ARMv6 support, rather than ARMv7.

This also adds support for crosstool-ng-based images, which are configured from template config files so that we can modify the GCC, glibc, and Linux versions. These are configured by running `crosstool-ng/configure.sh`, which scans that directory for all `*.config.in` files, and configures them to `docker/crosstool-config`. The script can have the toolchain versions modified via `GCC_VERSION`, `GLIBC_VERSION`, and `LINUX_VERSION`.

Closes #426.