Fix docker detection.
Currently, docker detection fails under a few cases:
1. If `podman_docker` is installed, so the filename is docker but the
actual executable is podman. 2. If the executable has a suffix, such as `.exe` on Windows, because we check if the executable `ends_with(DOCKER)`.

The only reliable way to fix both these issues, IE, if the actual engine is docker and not an alias, and if the executable does not contain a suffix, is to query the container engine. This might not be ideal for performance reasons, but is the only reliable way to fix these issues.

Emulated docker will have an output of `Emulate Docker CLI using podman.`, while docker itself will have an output similar to `Docker version 20.10.16, build aa7e414`.

Closes #530.
Closes #538.