Bug fix for dragonfly and CI tests.
Added `build-std` flag to CI, so for targets that support `std` but are
tier 3 targets, we can use the experimental `build-std` support to
ensure they build. Also patches linking on dragonfly to ensure we use
`libgcc_s` instead of `libgcc_pic`, which no longer exists. We just
create a symlink for that purpose.

This also adds `build-std` detection into Rust, so we don't have to
provide it manually. It allows use to avoid using `xargo` when possible.

Necessary to merge #775.