Add build-std config option.
Detects if `rust-std` is not an available component for a target, and if `build-std = true` in `Cross.toml` under `[build]` or `[target.(...)]`, then use the `-Zbuild-std` flag. This has higher priority than `xargo`: `xargo` will only be used if a target is not a built-in.

A sample `Cross.toml` file is the following:

```toml
[target.x86_64-unknown-dragonfly]
build-std = true
```

Closes #692.