Support temporary data directories and files.
Create and robustly delete temporary data, which is stored in the user data directory under `${data_dir}/cross-rs/tmp`. Therefore, if program execution is ever halted, deleting the directory will automatically cleanup any remaining data.

A termination handler cleans up the temporary data on exit by clearing a stack of temporary files/directories, and we have 2 resource handlers that cleanup the temporary data when the object is dropped.

The new installer for the startup hooks is `install_exit_hooks`, which installs both the panic and termination handlers. To get the directory containing all temporary data, use `cross::temp::dir()`.

An example of using temporary directories and files is:

```rust
{
    // these are all safe due to single-threaded execution
    let tmpdir1 = unsafe { temp::TempFile::new() }?;
    let tmpdir2 = unsafe { temp::TempFile::new() }?;
    let tmpfile1 = unsafe { temp::TempFile::new() }?;
    let tmpfile2 = unsafe { temp::TempFile::new() }?;

    for entry in fs::read_dir(tmpdir1.path()) {
        ...
    }
    for entry in fs::read_dir(tmpdir2.path()) {
        ...
    }
}   // cleanup tmpfile2 -> tmpfile1 -> tmpdir2 -> tmpdir1
```

Note that only these 2 wrappers are provided, since it guarantees the temporary file and directory stack stays ordered and resources are cleaned up as desired.