Avoid using display on paths.
Adds a helper trait `ToUtf8` which converts the ath to UTF-8 if
possible, and if not, creates a pretty debugging representation of the
path. We can then change `display()` to `to_utf8()?` and be completely
correct in all cases.

On POSIX systems, this doesn't matter since paths must be UTF-8 anyway,
but on Windows some paths can be WTF-8 (see @rust-lang/rust#12056).
Either way, this can avoid creating a copy in cases, is always correct,
and is more idiomatic about what we're doing.