Fixes regression in #828 with absolute path.
Prior to #828, we added a '/' separator before component in our
makeshift `as_posix`, so the path would always become absolute, even
though the actual path was relative. `as_posix` therefore handled this
as a relative path, so using `project` and not `/project` as the working
directory for the cross command caused commands to fail.