Add file fingerprint to copy/remove only required files.
Updates the persistent volume using a fingerprint of all files in the
project, skipping any cache directories by default. If the file modified
date has changed, or the file has been added, copy it to the volume and
update it. If the file has been removed, then remove it from the host.
To avoid a massive command-line argument, we copy a file containing each
changed file on a line to the container, and then remove each file
by running a script on the container.