Enable copying symlinks when using remote docker.
This handles the reads the link dst from the symlink when using
temporary directories, and creates a softlink to the path (keeping
relative and absolute symlinks valid) in the temporary directory, which
is then copied with the same permissions over the dst filesystem. This
should not keep the same metadata, but since we fingerprint source
files, this should not matter.