Add integration tests.
Adds an an integration test for remote Docker support with and without persistent data volumes, and adds it to CI. In addition, we've added integration tests for custom toolchains using `cargo-bisect-rustc`, and tests for docker-in-docker scenarios. Also changes the workflow  to ensure publish depends on fmt, clippy, and cargo-deny. Fixes a bug  when removing data volumes, since it does not require a target.

The docker-in-docker scenarios both run the cross unittests, and test workspace and manifest paths.