Always link to libc when compiling musl targets.
This fixes a regression when changing musl targets to always statically link to libstdc++, and the static version of libstdc++ can reference but be missing requiring symbols from libc. Due to the linking order of Rust, these symbols sometimes are not resolved. This ensures that all binaries with external C++ dependencies cross-compiled for musl targets, such as `re2`, will be properly linked.

`aarch64-unknown-linux-musl` has an additional issue, where it requires the compiler builtins to be linked at the end, since some of the symbols reference in the static `libstdc++` reference those provided by either compiler builtins or libgcc.