Bugfix: custom image tag can't end with neither a `-` nor `.`
The original bug fix solved the problem of a image name ending in a single `-` character (which docker doesn't permit). However, docker also doesn't permit names ending in a `.`, so we need to handle that case too.

Also, removing just the last char is insufficient, as names may end up ending with multiple consecutive `-` or `.` characters, so we need to loop until they're all gone.